+++ react/hooks/useFacetNavigation.js
@@ -175,7 +175,7 @@ export const buildNewQueryMap = (
   return buildQueryAndMap(querySegments, mapSegments, facets, selectedFacets)
 }
 
-const useFacetNavigation = selectedFacets => {
+const useFacetNavigation = (selectedFacets, scrollToTop = 'none') => {
   const { navigate, setQuery } = useRuntime()
   const { map, query } = useFilterNavigator()
   const { fuzzy, operator, searchState } = useSearchState()
@@ -190,6 +190,10 @@ const useFacetNavigation = selectedFacets => {
         facets,
         selectedFacets
       )
+
+      if (scrollToTop !== 'none') {
+        window.scroll({ top: 0, left: 0, behavior: scrollToTop })
+      }
 
       if (preventRouteChange) {
         const queries = {