+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Prop `scrollToTop` to `filter-navigator.v3` that scrolls the page to the top when selecting a facet.
 
 ## [3.68.1] - 2020-08-13
 ### Fixed